requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
passlib
python-jose[cryptography]